+++ VDA-DID-Registry/contracts/VeridaDIDRegistry.sol
@@ -5,8 +5,9 @@ import { VeridaDataVerificationLib } from "./VeridaDataVerificationLib.sol";
 import "@openzeppelin/contracts-upgradeable/access/OwnableUpgradeable.sol";
 import "@openzeppelin/contracts-upgradeable/utils/structs/EnumerableSetUpgradeable.sol";
 
+// Verida contract dependencies
+import { EnumerableSet } from "@verida/common-contract/contracts/EnumerableSet.sol";
 import "./IVeridaDIDRegistry.sol";
-import "./EnumerableSet.sol";
 
 // import "hardhat/console.sol";