+++ VDA-Name-Registry/hardhat.config.ts
@@ -10,13 +10,7 @@ import "@openzeppelin/hardhat-upgrades";
 import "@openzeppelin/hardhat-defender";
 
 dotenv.config({ path: __dirname + "/.env" });
-const {
-  PRIVATE_KEY,
-  ETHERSCAN_API_KEY,
-  POLYGONSCAN_API_KEY,
-  RPC_URL_POLYGON,
-  RPC_URL_MUMBAI,
-} = process.env;
+const {PRIVATE_KEY, ETHERSCAN_API_KEY, POLYGONSCAN_API_KEY, POLYGON_TESTNET_RPC, POLYGON_MAINNET_RPC} = process.env;
 
 const config: HardhatUserConfig = {
   defender: {
@@ -38,14 +32,14 @@ const config: HardhatUserConfig = {
   },
   networks: {
     polygonmainnet: {
-      url: RPC_URL_POLYGON !== undefined? RPC_URL_POLYGON : "https://polygon-rpc.com/",
+      url: POLYGON_MAINNET_RPC !== undefined ? POLYGON_MAINNET_RPC : "https://polygon-rpc.com/",
       chainId: 137,
-      accounts: [PRIVATE_KEY!],
+      accounts: PRIVATE_KEY !== undefined ? [PRIVATE_KEY] : [],
     },
     polygontestnet: {
-      url: RPC_URL_MUMBAI !== undefined? RPC_URL_MUMBAI : "https://rpc-mumbai.maticvigil.com",
+      url: POLYGON_TESTNET_RPC !== undefined ? POLYGON_TESTNET_RPC : "https://matic-mumbai.chainstacklabs.com",
       chainId: 80001,
-      accounts: [PRIVATE_KEY!],
+      accounts: PRIVATE_KEY !== undefined ? [PRIVATE_KEY] : [], 
     },
     goerli: {
       url: "https://eth-goerli.public.blastapi.io", //https://goerli.infura.io/v3/