+++ VDA-SBT/hardhat.config.ts
@@ -1,21 +1,21 @@
-import * as dotenv from "dotenv";
-
 import { HardhatUserConfig } from "hardhat/config";
 import "@nomicfoundation/hardhat-toolbox";
 
-// For BSC verification after deploy
+import * as dotenv from "dotenv";
+
+// For verify
 import "@nomiclabs/hardhat-ethers";
 // For upgradeable - deploy
 import "@openzeppelin/hardhat-upgrades";
 
-dotenv.config({path: __dirname + '/.env'});
-const {PRIVATE_KEY, BSCSCAN_API_KEY, POLYGONSCAN_API_KEY, POLYGON_TESTNET_RPC, POLYGON_MAINNET_RPC} = process.env;
+dotenv.config({path: __dirname + '/.env'})
+const {PRIVATE_KEY, POLYGONSCAN_API_KEY, POLYGON_TESTNET_RPC, POLYGON_MAINNET_RPC} = process.env;
 
 const config: HardhatUserConfig = {
   solidity: {
     compilers: [
       {
-        version: "0.8.17",
+        version: "0.8.18",
         settings: {
           optimizer: {
             enabled: true,
@@ -32,17 +32,6 @@ const config: HardhatUserConfig = {
     artifacts: "./artifacts",
   },
   networks: {
-    bsctestnet: {
-      url: "https://data-seed-prebsc-1-s1.binance.org:8545/",
-      // url: "https://speedy-nodes-nyc.moralis.io/bd1c39d7c8ee1229b16b4a97/bsc/testnet",
-      chainId: 97,
-      accounts: PRIVATE_KEY !== undefined ? [PRIVATE_KEY] : [], 
-    },
-    bscmainnet: {
-      url: "https://bsc-dataseed.binance.org/",
-      chainId: 56,
-      accounts: PRIVATE_KEY !== undefined ? [PRIVATE_KEY] : [], 
-    },
     polygonmainnet: {
       url: POLYGON_MAINNET_RPC !== undefined ? POLYGON_MAINNET_RPC : "https://polygon-rpc.com/",
       chainId: 137,