+++ VDA-Verification-Base/hardhat.config.ts
@@ -3,20 +3,19 @@ import "@nomicfoundation/hardhat-toolbox";
 
 import * as dotenv from "dotenv";
 
-// For Verify 
+// For verify
 import "@nomiclabs/hardhat-ethers";
 // For upgradeable - deploy
 import "@openzeppelin/hardhat-upgrades";
 
-dotenv.config()
-
-const {PRIVATE_KEY, POLYGONSCAN_API_KEY} = process.env;
+dotenv.config({path: __dirname + '/.env'})
+const {PRIVATE_KEY, POLYGONSCAN_API_KEY, POLYGON_TESTNET_RPC, POLYGON_MAINNET_RPC} = process.env;
 
 const config: HardhatUserConfig = {
   solidity: {
     compilers: [
       {
-        version: "0.8.17",
+        version: "0.8.18",
         settings: {
           optimizer: {
             enabled: true,
@@ -34,15 +33,15 @@ const config: HardhatUserConfig = {
   },
   networks: {
     polygonmainnet: {
-      url: "https://polygon-rpc.com/",
+      url: POLYGON_MAINNET_RPC !== undefined ? POLYGON_MAINNET_RPC : "https://polygon-rpc.com/",
       chainId: 137,
       accounts: PRIVATE_KEY !== undefined ? [PRIVATE_KEY] : [],
     },
     polygontestnet: {
-      url: "http://44.234.36.28:8545",
+      url: POLYGON_TESTNET_RPC !== undefined ? POLYGON_TESTNET_RPC : "https://matic-mumbai.chainstacklabs.com",
       chainId: 80001,
       accounts: PRIVATE_KEY !== undefined ? [PRIVATE_KEY] : [], 
-    }
+    },
   },
   gasReporter: {
     enabled: process.env.REPORT_GAS !== undefined,